requirements.txt
hyperliquid-python-sdk>=0.2.0
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=12.0
